Mark the capbility flag <iOMXComponentSupportsMovableInputBuffers> as FALSE within PSB component.
For codec component plugins used by OpenCORE, there are two input buffer
usage models that could be selected via component's capbility setting
flag.

When iOMXComponentSupportsMovableInputBuffer = TRUE, input data is
contained within a media data shared pointer and will be delivered
directly from the upstream component to the codec component without
extra copying. The actual buffer used within ProcessorProcess() is not
the OMX-IL defined input buffers, but the media data shared pointer.
In this case, only the OMX_BUFFERHEADERTYPE->pBuffer and nFilledLen is
updated by OpenCORE and however the nAllocLen is not equal to the buffer
size for the shared pointer. Hence, the psb component must guarantee not
touch data bits beyond the pBuffer[0]~pBuffer[nFilledLen-1]. Currently,
the MI-X doesn't guarantee this (at least within
mix_videoconfigparamsdec_set_header) and will cause memroy corruption
sometimes.

When iOMXComponentSupportsMovableInputBuffer = FALSE, input data is
contained within the OMX-IL defined input buffers. The actual buffer
used within ProcessorProcess() is definately the buffer defined by
OMX-IL. Hence, OMX_BUFFERHEADERTYPE->pBuffer, nFilledLen, nAllocLen are
all ok. It is safe for the psb component to touch any bits with the
scope from pBuffer[0] ~ pBuffer[nAllocLen - 1];

For the MI-X issue, the iOMXComponentSupportsMovableInputBuffer is set
as FALSE. There will be an extra bitstream data copy between the
upstream component and the codec component and the performance might
suffer. However, the memory leak risk is eilmiated.

There are way for improving the component efficiency when guaranteeing
the memroy leak risk is zero, that is, to mark the MixIOVec's
buffer_size as the OMX_BUFFERHEADERTYPE->nFilledLen instead of
nAllocLen. However, this is an OpenCORE specific fix and need to be
marked. (Will be carried out when optimizing)

Change-Id: I3affe4f490fedbb1cd17b09eb852f75a6adb1195
Signed-off-by: Liu, Shuo <shuo.liu@intel.com>