videocodec: New mechanism to prevent deadlock on trick play
A few changes to better handle cessation of video decode
during trick play and to prevent a deadlock.  We should
be flushing the output ports. On detection of reporting
no surface during processing mark indexes for retain
and reset fill length.  Secondly, implement
ProcessorReleaseLock() with the video decoder's own
release lock.  This virtual method is handled in the
component base of the core.

Change-Id: I04b914425edd18401724ea0039be4250318476f9
Depends-Change-Id: I26e214178d6e7106a1a9d4b5086eb8ba8e3a4913
Depends-Change-Id: Ibca0f67e07d6ce2391711a72cb8466f0dcab5fa8
Signed-off-by: Robin Mathew <robbin.mathew@intel.com>
Signed-off-by: Sean V Kelley <sean.v.kelley@intel.com>