videodecoder: fix buffer mapping logic to remove memory leak
This patch fixes the following things:

* function MapRawNV12 modifies the nFilledLen variable when
called, so the nFilledLen assignation was moved up to reflect when
a NativeWindow buffer is used.

* function MapRawNV12 uses Surfaces which internally contain vaImage.
There's no need to create new vaImages to map data to user space, instead
the vaImage is derived from the current surface.  This allows re-usability
of vaImages and reduces potential memory leak

* function MapRawNV12 needs to call vaUnmapBuffer once it has modified its
content.  In this case the content is copied to a OMX buffer so it is a must
to unmap and later destroy the Image derived per libva API.

* after calling MapRaw12 and when NativeWindow buffer is not user the
surface can be set so that it can be re-used as fast as possible.

Both thumbnail generation and video playback benefit from this change.

For: AXIA-328

Change-Id: I77f025f9aaabfe8db37b960c04b9e7ee73670e0c
Signed-off-by: Daniel Charles <daniel.charles@intel.com>