omx-components: Load shared libraries plug-ins on initialization
Sandboxing on Chromium OS requires that libraries are loaded
at the creation of objects before it is locked down.  The
libraries to be used by the plug-in hierarchy are made available
and are passed down to the parser.

Due to the locking of sandbox, instead of having the configuration
file to add components, the components are loaded from a table
defined within the header file.

Change-Id: I207754912ed5a3a3670fe72e95c4c64e8307b928
Signed-off-by: Daniel Charles <daniel.charles@intel.com>