omx-videdecoder: add support for rendering images
On a X11 environment, the il-client can provide the images or
pixmaps to use to render the decoded information.  The pixmaps can be
used to call libva API to have the information properly rendered. This
way the il-client does not need to link against libva and the omx-il
implementation does not need to allocate GLX specific stuff that belongs
to the il-client.

Change-Id: I7f3bec9773e293582b81554a63d3b0f95bf4cfae
Signed-off-by: Daniel Charles <daniel.charles@intel.com>