omx-viddec: workaround to keep omx and client buffer count
The whole else case on FillRenderBuffer function has to be re-written.
The omx_buffer reference kept inside the RenderBuffer should not be
used to change the buffer to be returned.  ProcessorProcess it the one
method that should return the buffers.

By avoiding the multiple return of the same buffer when the omx_buffer
reference is not kept inside RenderBuffer, the buffer count at the client
keeps within valid values.

Change-Id: Id01c155214b9c004dfd8f74d81046fbe5732adaa
Signed-off-by: Daniel Charles <daniel.charles@intel.com>