Resolution change handling support for VP8 OMX.
Added code to handle resolution change for VP8 OMX
component. The H.264 OMX component performs a flush when
new resolution is detected though SPS header. However,
a flush will make us lose the key frame in case of VP8.
This causes all subsequent inter-frames to be dropped
due to missing reference. The current solution is to not
do a flush in VP8 case, but simply ignore the previous output
buffer with incorrect resolution. This works for VP8 IVF and WebM
files. Dynamic resolution change handling with flush will be added
later.

Change-Id: I3339d00f000504dc95153a26c87579bd9300968b
Signed-off-by: Sameer Kibey <sameer.kibey@intel.com>